Only mark personal fields mandatory when type dropdown is set to Personal
The problem was that after setting a contributor to Organisational and filling in the name and isni the check on line 72 failed and the check on 75 passed and since family name and given name are present but empty it went on to execute 78.  This change skips the check on 75 as it should when the type is Organisational.